moved dragover directive
