+++ projects/limble-tree/src/lib/components/dropzone/dropzone.component.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from "@angular/common";
 import { Component, EventEmitter, Input, Output } from "@angular/core";
-import { DragoverNoChangeDetectDirective } from "../branch/dragover-no-change-detect";
+import { DragoverNoChangeDetectDirective } from "../../extras/drag-and-drop/dragover-no-change-detect";
 
 @Component({
    standalone: true,
