+++ projects/limble-tree/src/lib/extras/drag-and-drop/index.ts
@@ -1,3 +1,4 @@
 export * from "./drag-and-drop.module";
 export * from "./drag-and-drop.service";
 export * from "./draggable.directive";
+export * from "./dragover-no-change-detect";
